+++ src/main/java/com/sosim/server/jwt/JwtFactory.java
@@ -4,6 +4,6 @@ package com.sosim.server.jwt;
  * <h2>로그인 인증 토큰은 생성하는 객체</h2>
  */
 public interface JwtFactory {
-    String createAccessToken(String id, String email);
+    String createAccessToken(String id);
     String createRefreshToken();
 }
