+++ src/main/java/com/sosim/server/jwt/JwtProvider.java
@@ -2,7 +2,6 @@ package com.sosim.server.jwt;
 
 import java.util.Optional;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 public interface JwtProvider {
 
@@ -11,10 +10,8 @@ public interface JwtProvider {
     Optional<String> extractId(String AccessToken);
     boolean isTokenValid(String token);
     String reIssueRefreshToken(String id);
-    String reIssueRefreshToken(RefreshToken refreshToken);
+//    String reIssueRefreshToken(RefreshToken refreshToken);
     String verifyAccessToken(String accessToken);
     String refresh(String refreshToken);
-    void setAccessTokenHeader(HttpServletResponse response, String accessToken);
-    void setRefreshTokenHeader(HttpServletResponse response, String refreshToken);
 
 }
