+++ src/main/java/com/sosim/server/jwt/RefreshToken.java
@@ -19,8 +19,7 @@ public class RefreshToken {
     // User
     private String id;
 
-    private String socialType;
-
-    private String socialId;
+//    private String socialType;
+//    private String socialId;
 
 }
